vlms-hallucinate/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.34.0
    # via -r requirements.in
anyio==4.4.0
    # via
    #   anthropic
    #   httpx
    #   openai
cachetools==5.4.0
    # via google-auth
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
distro==1.9.0
    # via
    #   anthropic
    #   openai
filelock==3.15.4
    # via huggingface-hub
fsspec==2024.6.1
    # via huggingface-hub
google-ai-generativelanguage==0.6.6
    # via google-generativeai
google-api-core==2.19.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.141.0
    # via google-generativeai
google-auth==2.33.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.7.2
    # via -r requirements.in
googleapis-common-protos==1.63.2
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.65.4
    # via
    #   google-api-core
    #   grpcio-status
    #   grpcio-tools
grpcio-status==1.62.3
    # via google-api-core
grpcio-tools==1.62.3
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.27.0
    # via
    #   anthropic
    #   openai
huggingface-hub==0.24.5
    # via tokenizers
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.5.0
    # via
    #   anthropic
    #   openai
openai==1.40.7
    # via -r requirements.in
packaging==24.1
    # via huggingface-hub
parameterized==0.9.0
    # via -r requirements.in
pillow==10.4.0
    # via -r requirements.in
proto-plus==1.24.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==4.25.4
    # via
    #   -r requirements.in
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   grpcio-tools
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.8.2
    # via
    #   anthropic
    #   google-generativeai
    #   openai
pydantic-core==2.20.1
    # via pydantic
pymupdf==1.24.9
    # via -r requirements.in
pymupdfb==1.24.9
    # via pymupdf
pyparsing==3.1.2
    # via httplib2
pyyaml==6.0.2
    # via huggingface-hub
requests==2.32.3
    # via
    #   google-api-core
    #   huggingface-hub
rsa==4.9
    # via google-auth
setuptools==72.2.0
    # via grpcio-tools
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
    #   openai
tabulate==0.9.0
    # via -r requirements.in
tokenizers==0.20.0
    # via anthropic
tqdm==4.66.5
    # via
    #   google-generativeai
    #   huggingface-hub
    #   openai
typing-extensions==4.12.2
    # via
    #   anthropic
    #   google-generativeai
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.2
    # via requests
